feat(nextbike/visualization): add file only for folium plots + fix moment plot
